Move user mute list item components
